meterData/requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
pika==1.3.2
protobuf==5.27.1
